+++ client-app/src/examples/portfolio/detail/charts/LineChart.ts
@@ -3,10 +3,14 @@ import {creates, hoistCmp, HoistModel, lookup, managed, XH} from '@xh/hoist/core
 import {fmtDate} from '@xh/hoist/format';
 import {ChartsPanelModel} from './ChartsPanelModel';
 import {panel} from '@xh/hoist/desktop/cmp/panel';
+import {errorMessage} from '@xh/hoist/dynamics/desktop';
 
 export const lineChart = hoistCmp.factory({
     model: creates(() => LineChartModel),
-    render() {
+    render({model}) {
+        if (model.lastLoadException) {
+            return errorMessage({error: model.lastLoadException});
+        }
         return panel({
             item: chart(),
             mask: 'onLoad',
@@ -70,13 +74,14 @@ class LineChartModel extends HoistModel {
 
         try {
             const series = await XH.portfolioService.getLineChartSeriesAsync({symbol, loadSpec});
-
-            if (!loadSpec.isObsolete) {
+            if (!loadSpec.isStale) {
                 chartModel.setSeries(series);
             }
         } catch (e) {
+            if (loadSpec.isAutoRefresh || loadSpec.isStale) return;
             chartModel.clear();
-            XH.handleException(e);
+            XH.handleException(e, {showAlert: false});
+            throw e;
         }
     }
 }