+++ grails-app/init/io/xh/toolbox/BootStrap.groovy
@@ -38,9 +38,8 @@ class BootStrap {
     //------------------------
     @Transactional
     private void createLocalAdminUserIfNeeded() {
-        String adminUsername = getInstanceConfig('adminUsername')
-        String adminPassword = getInstanceConfig('adminPassword')
-
+        String adminUsername = getInstanceConfig('bootstrapAdminUser')
+        String adminPassword = getInstanceConfig('bootstrapAdminPassword')
         if (adminUsername && adminPassword) {
             def user = User.findByEmail(adminUsername)
             if (!user) {
@@ -70,13 +69,27 @@ class BootStrap {
    \\ \\_\\  \\ \\_____\\  \\ \\_____\\  \\ \\_____\\  \\ \\_____\\  \\ \\_____\\   /\\_\\/\\_\\ 
     \\/_/   \\/_____/   \\/_____/   \\/_____/   \\/_____/   \\/_____/   \\/_/\\/_/ 
 \n                                                                           
-         ${appName} v${appVersion} - ${appEnvironment}
+         ${appName} v${appVersion} [build ${appBuild}] - ${appEnvironment}
 \n
         """)
     }
 
     private void ensureRequiredConfigsCreated() {
         configService.ensureRequiredConfigsCreated([
+            auth0ClientId: [
+                    valueType: 'string',
+                    defaultValue: 'MUn9VrAGavF7n39RdhFYq8xkZkoFYEDB',
+                    clientVisible: false,
+                    groupName: 'Auth0',
+                    note: 'Client ID of the Toolbox app registered at our Auth0 account. \n(https://manage.auth0.com/dashboard/us/xhio/)'
+            ],
+            auth0Domain: [
+                    valueType: 'string',
+                    defaultValue: 'login.xh.io',
+                    clientVisible: false,
+                    groupName: 'Auth0',
+                    note: 'Custom domain for our Auth0 deployment. OAuth login flow will redirect users here.'
+            ],
             contacts: [
                     valueType: 'json',
                     defaultValue: [