+++ CHANGELOG.md
@@ -2,7 +2,17 @@
 
 ## v4.0-SNAPSHOT - unreleased
 
+### New Features
+* Adds new `bootstrapLogin` app that uses `bootstrapAdminUser` and `bootstrapAdminPassword`to bypass
+`Auth0` login in development mode when `Auth0` itself isn't working or when running the app via 
+`yarn startOnDevice`.
+
+### Technical
+* Remove obsolete `useOAuth` and `adminUsername` instanceConfigs.  Simplify bootstrapping 
+process to always use `Auth0` with `bootstrapAdminUser` for initial login.
+
 ## v3.6.0 - 2024-01-19
+
 
 ### Libraries